cosmic_text

Enum Motion

pub enum Motion {

    LayoutCursor(LayoutCursor),
    Previous,
    Next,
    Left,
    Right,
    Up,
    Down,
    Home,
    SoftHome,
    End,
    ParagraphStart,
    ParagraphEnd,
    PageUp,
    PageDown,
    Vertical(i32),
    PreviousWord,
    NextWord,
    LeftWord,
    RightWord,
    BufferStart,
    BufferEnd,
    GotoLine(usize),

}

A motion to perform on a Cursor

Variants

LayoutCursor(LayoutCursor)

Apply specific LayoutCursor

Previous

Move cursor to previous character (Self::Left in LTR, Self::Right in RTL)

Next

Move cursor to next character (Self::Right in LTR, Self::Left in RTL)

Left

Move cursor left

Right

Move cursor right

Up

Move cursor up

Down

Move cursor down

Home

Move cursor to start of line

SoftHome

Move cursor to start of line, skipping whitespace

End

Move cursor to end of line

ParagraphStart

Move cursor to start of paragraph

ParagraphEnd

Move cursor to end of paragraph

PageUp

Move cursor up one page

PageDown

Move cursor down one page

Vertical(i32)

Move cursor up or down by a number of pixels

PreviousWord

Move cursor to previous word boundary

NextWord

Move cursor to next word boundary

LeftWord

Move cursor to next word boundary to the left

RightWord

Move cursor to next word boundary to the right

BufferStart

Move cursor to the start of the document

BufferEnd

Move cursor to the end of the document

GotoLine(usize)

Move cursor to specific line