cosmic_text

Struct FontSystem

pub struct FontSystem { /* private fields */ }

Access to the system fonts.

Implementations

impl FontSystem

pub fn new() -> Self

Create a new FontSystem, that allows access to any installed system fonts

Timing

This function takes some time to run. On the release build, it can take up to a second, while debug builds can take up to ten times longer. For this reason, it should only be called once, and the resulting FontSystem should be shared.

pub fn new_with_fonts(fonts: impl IntoIterator<Item = Source>) -> Self

Create a new FontSystem with a pre-specified set of fonts.

pub fn new_with_locale_and_db_and_fallback( locale: String, db: Database, impl_fallback: impl Fallback + 'static, ) -> Self

Create a new FontSystem with a pre-specified locale, font database and font fallback list.

pub fn new_with_locale_and_db(locale: String, db: Database) -> Self

Create a new FontSystem with a pre-specified locale and font database.

pub fn locale(&self) -> &str

Get the locale.

pub const fn db(&self) -> &Database

Get the database.

pub fn db_mut(&mut self) -> &mut Database

Get a mutable reference to the database.

pub fn into_locale_and_db(self) -> (String, Database)

Consume this FontSystem and return the locale and database.

pub fn get_font(&mut self, id: ID, weight: Weight) -> Option<Arc<Font>>

Get a font by its ID and weight.

pub fn is_monospace(&self, id: ID) -> bool

pub fn get_monospace_ids_for_scripts( &self, scripts: impl Iterator<Item = [u8; 4]>, ) -> Vec<ID>

pub fn get_font_supported_codepoints_in_word( &mut self, id: ID, weight: Weight, word: &str, ) -> Option<usize>

pub fn get_font_matches(&mut self, attrs: &Attrs<'_>) -> Arc<Vec<FontMatchKey>>