Struct cosmic_text::TextBuffer

pub struct TextBuffer<'a> {
    pub lines: Vec<TextBufferLine<'a>>,
    pub cursor_moved: bool,
    pub redraw: bool,
    /* private fields */
}

A buffer of text that is shaped and laid out

Fields
lines: Vec<TextBufferLine<'a>>

Lines (or paragraphs) of text in the buffer

cursor_moved: bool

True if the cursor has been moved. Set to false after processing

Usually, if this is true, you should run Self::shape_until_cursor before redrawing. Otherwise, you should run Self::shape_until_scroll

redraw: bool

True if a redraw is requires. Set to false after processing

Implementations

impl<'a> TextBuffer<'a>

pub fn new(font_system: &'a FontSystem<'a>, metrics: TextMetrics) -> Self

pub fn shape_until(&mut self, lines: i32) -> i32

Pre-shape lines in the buffer, up to lines, return actual number of layout lines

pub fn shape_until_cursor(&mut self)

Shape lines until cursor, also scrolling to include cursor in view

pub fn shape_until_scroll(&mut self)

Shape lines until scroll

pub fn cursor(&self) -> TextCursor

Get the current cursor position

pub fn metrics(&self) -> TextMetrics

Get the current TextMetrics

pub fn set_metrics(&mut self, metrics: TextMetrics)

Set the current TextMetrics

pub fn size(&self) -> (i32, i32)

Get the current buffer dimensions (width, height)

pub fn set_size(&mut self, width: i32, height: i32)

Set the current buffer dimensions

pub fn scroll(&self) -> i32

Get the current scroll location

pub fn lines(&self) -> i32

Get the number of lines that can be viewed in the buffer

pub fn set_text(&mut self, text: &str, attrs: Attrs<'a>)

Set text of buffer, using provided attributes for each line by default

pub fn action(&mut self, action: TextAction)

Perform a TextAction on the buffer

pub fn hit(&self, x: i32, y: i32) -> Option<TextCursor>

Convert x, y position to TextCursor (hit detection)

pub fn layout_runs<'b>(&'b self) -> TextLayoutRunIter<'a, 'b>

Get the visible layout runs for rendering and other tasks

pub fn draw<F>(&self, cache: &mut SwashCache<'_>, color: Color, f: F) where
F: FnMut(i32, i32, u32, u32, Color),

Draw the buffer