Crate cosmic_text

COSMIC Text

This library provides advanced text handling in a generic way. It provides abstractions for shaping, font discovery, font fallback, layout, rasterization, and editing. Shaping utilizes rustybuzz, font discovery utilizes fontdb, and the rasterization is optional and utilizes swash. The other features are developed internal to this library.

It is recommended that you start by creating a FontSystem, after which you can create a Buffer, provide it with some text, and then inspect the layout it produces. At this point, you can use the SwashCache to rasterize glyphs into either images or pixels.

use cosmic_text::{Attrs, Color, FontSystem, SwashCache, Buffer, Metrics};

// A FontSystem provides access to detected system fonts, create one per application
let font_system = FontSystem::new();

// A SwashCache stores rasterized glyphs, create one per application
let mut swash_cache = SwashCache::new(&font_system);

// Text metrics indicate the font size and line height of a buffer
let metrics = Metrics::new(14, 20);

// A Buffer provides shaping and layout for a UTF-8 string, create one per text widget
let mut buffer = Buffer::new(&font_system, metrics);

// Set a size for the text buffer, in pixels
buffer.set_size(80, 25);

// Attributes indicate what font to choose
let attrs = Attrs::new();

// Add some text!
buffer.set_text("Hello, Rust! 🦀\n", attrs);

// Perform shaping as desired
buffer.shape_until_scroll();

// Inspect the output runs
for run in buffer.layout_runs() {
    for glyph in run.glyphs.iter() {
        println!("{:#?}", glyph);
    }
}

// Create a default text color
let text_color = Color::rgb(0xFF, 0xFF, 0xFF);

// Draw the buffer (for performance, instead use SwashCache directly)
buffer.draw(&mut swash_cache, text_color, |x, y, w, h, color| {
    // Fill in your code here for drawing rectangles
});

Structs

Attrs

Text attributes

AttrsList

List of text attributes to apply to a line

AttrsOwned

Buffer

A buffer of text that is shaped and laid out

BufferLine

A line (or paragraph) of text that is shaped and laid out

CacheKey

Key for building a glyph cache

Color

Text color

Cursor

Current cursor location

Editor

A wrapper of Buffer for easy editing

FontMatches

Fonts that match a pattern

FontSystem

Access system fonts

LayoutCursor

LayoutGlyph

A laid out glyph

LayoutLine

A line of laid out glyphs

LayoutRun

A line of visible text for rendering

LayoutRunIter

An iterator of visible text lines, see LayoutRun

Metrics

Metrics of text

ShapeGlyph

A shaped glyph

ShapeLine

A shaped line (or paragraph)

ShapeSpan

A shaped span (for bidirectional processing)

ShapeWord

A shaped word (for word wrapping)

SwashCache

Cache for rasterizing with the swash scaler

SwashImage

Scaled glyph image.

Weight

Specifies the weight of glyphs in the font, their degree of blackness or stroke thickness.

Enums

Action

An action to perform on an Editor

Family

A font family.

FamilyOwned

Stretch

A face width.

Style

Allows italic or oblique faces to be selected.

SubpixelBin

Binning of subpixel position for cache optimization

SwashContent

Content of a scaled glyph image.