Struct cosmic_text::BorrowedWithFontSystem

pub struct BorrowedWithFontSystem<'a, T> { /* private fields */ }

A value borrowed together with an FontSystem

Implementations

impl<'a> BorrowedWithFontSystem<'a, Buffer>

pub fn shape_until(&mut self, lines: i32) -> i32

Pre-shape lines in the buffer, up to lines, return actual number of layout lines

pub fn shape_until_cursor(&mut self, cursor: Cursor)

Shape lines until cursor, also scrolling to include cursor in view

pub fn shape_until_scroll(&mut self)

Shape lines until scroll

pub fn line_shape(&mut self, line_i: usize) -> Option<&ShapeLine>

Shape the provided line index and return the result

pub fn line_layout(&mut self, line_i: usize) -> Option<&[LayoutLine]>

Lay out the provided line index and return the result

pub fn set_metrics(&mut self, metrics: Metrics)

Set the current Metrics

Panics

Will panic if metrics.font_size is zero.

pub fn set_wrap(&mut self, wrap: Wrap)

Set the current Wrap

pub fn set_size(&mut self, width: f32, height: f32)

Set the current buffer dimensions

pub fn set_text(&mut self, text: &str, attrs: Attrs<'_>)

Set text of buffer, using provided attributes for each line by default

pub fn draw<F>(&mut self, cache: &mut SwashCache, color: Color, f: F)where F: FnMut(i32, i32, u32, u32, Color),

Draw the buffer

impl<'a, T: Edit> BorrowedWithFontSystem<'a, T>

pub fn buffer_mut(&mut self) -> BorrowedWithFontSystem<'_, Buffer>

Get the internal Buffer, mutably

pub fn shape_as_needed(&mut self)

Shape lines until scroll, after adjusting scroll if the cursor moved

pub fn action(&mut self, action: Action)

Perform an Action on the editor

pub fn draw<F>(&mut self, cache: &mut SwashCache, color: Color, f: F)where F: FnMut(i32, i32, u32, u32, Color),

Draw the editor