Struct cosmic_text::BufferLine

pub struct BufferLine { /* private fields */ }

A line (or paragraph) of text that is shaped and laid out

Implementations

impl BufferLine

pub fn new<T: Into<String>>(text: T, attrs_list: AttrsList) -> Self

Create a new line with the given text and attributes list Cached shaping and layout can be done using the Self::shape and Self::layout functions

pub fn text(&self) -> &str

Get current text

pub fn set_text<T: AsRef<str> + Into<String>>(
&mut self,
text: T,
attrs_list: AttrsList
) -> bool

Set text and attributes list

Will reset shape and layout if it differs from current text and attributes list. Returns true if the line was reset

pub fn attrs_list(&self) -> &AttrsList

Get attributes list

pub fn set_attrs_list(&mut self, attrs_list: AttrsList) -> bool

Set attributes list

Will reset shape and layout if it differs from current attributes list. Returns true if the line was reset

pub fn wrap_simple(&self) -> bool

Get simple wrapping setting (wrap by characters only)

pub fn set_wrap_simple(&mut self, wrap_simple: bool) -> bool

Set simple wrapping setting (wrap by characters only)

Will reset shape and layout if it differs from current simple wrapping setting. Returns true if the line was reset

pub fn append(&mut self, other: Self)

Append line at end of this line

The wrap setting of the appended line will be lost

pub fn split_off(&mut self, index: usize) -> Self

Split off new line at index

pub fn reset(&mut self)

Reset shaping and layout information

pub fn reset_layout(&mut self)

Reset only layout information

pub fn is_reset(&self) -> bool

Check if shaping and layout information is cleared

pub fn shape(&mut self, font_system: &FontSystem) -> &ShapeLine

Shape line, will cache results

pub fn shape_opt(&self) -> &Option<ShapeLine>

Get line shaping cache

pub fn layout(
&mut self,
font_system: &FontSystem,
font_size: i32,
width: i32
) -> &[LayoutLine]

Layout line, will cache results

pub fn layout_opt(&self) -> &Option<Vec<LayoutLine>>

Get line layout cache