Struct cosmic_text::BufferLine

pub struct BufferLine { /* private fields */ }

A line (or paragraph) of text that is shaped and laid out

Implementations

impl BufferLine

pub fn new<T: Into<String>>( text: T, ending: LineEnding, attrs_list: AttrsList, shaping: Shaping, ) -> Self

Create a new line with the given text and attributes list Cached shaping and layout can be done using the Self::shape and Self::layout functions

pub fn reset_new<T: Into<String>>( &mut self, text: T, ending: LineEnding, attrs_list: AttrsList, shaping: Shaping, )

Resets the current line with new internal values.

Avoids deallocating internal caches so they can be reused.

pub fn text(&self) -> &str

Get current text

pub fn set_text<T: AsRef<str>>( &mut self, text: T, ending: LineEnding, attrs_list: AttrsList, ) -> bool

Set text and attributes list

Will reset shape and layout if it differs from current text and attributes list. Returns true if the line was reset

pub fn into_text(self) -> String

Consume this line, returning only its text contents as a String.

pub fn ending(&self) -> LineEnding

Get line ending

pub fn set_ending(&mut self, ending: LineEnding) -> bool

Set line ending

Will reset shape and layout if it differs from current line ending. Returns true if the line was reset

pub fn attrs_list(&self) -> &AttrsList

Get attributes list

pub fn set_attrs_list(&mut self, attrs_list: AttrsList) -> bool

Set attributes list

Will reset shape and layout if it differs from current attributes list. Returns true if the line was reset

pub fn align(&self) -> Option<Align>

Get the Text alignment

pub fn set_align(&mut self, align: Option<Align>) -> bool

Set the text alignment

Will reset layout if it differs from current alignment. Setting to None will use Align::Right for RTL lines, and Align::Left for LTR lines. Returns true if the line was reset

pub fn append(&mut self, other: Self)

Append line at end of this line

The wrap setting of the appended line will be lost

pub fn split_off(&mut self, index: usize) -> Self

Split off new line at index

pub fn reset(&mut self)

Reset shaping, layout, and metadata caches

pub fn reset_shaping(&mut self)

Reset shaping and layout caches

pub fn reset_layout(&mut self)

Reset only layout cache

pub fn shape( &mut self, font_system: &mut FontSystem, tab_width: u16, ) -> &ShapeLine

Shape line, will cache results

pub fn shape_opt(&self) -> Option<&ShapeLine>

Get line shaping cache

pub fn layout( &mut self, font_system: &mut FontSystem, font_size: f32, width_opt: Option<f32>, wrap: Wrap, match_mono_width: Option<f32>, tab_width: u16, ) -> &[LayoutLine]

Layout line, will cache results

pub fn layout_opt(&self) -> Option<&Vec<LayoutLine>>

Get line layout cache

pub fn metadata(&self) -> Option<usize>

Get line metadata. This will be None if BufferLine::set_metadata has not been called after the last reset of shaping and layout caches

pub fn set_metadata(&mut self, metadata: usize)

Set line metadata. This is stored until the next line reset