Trait cosmic_text::Edit

pub trait Edit {

    // Required methods
    fn buffer(&self) -> &Buffer;
    fn buffer_mut(&mut self) -> &mut Buffer;
    fn cursor(&self) -> Cursor;
    fn set_cursor(&mut self, cursor: Cursor);
    fn select_opt(&self) -> Option<Cursor>;
    fn set_select_opt(&mut self, select_opt: Option<Cursor>);
    fn shape_as_needed(&mut self, font_system: &mut FontSystem);
    fn copy_selection(&mut self) -> Option<String>;
    fn delete_selection(&mut self) -> bool;
    fn insert_string(&mut self, data: &str, attrs_list: Option<AttrsList>);
    fn action(&mut self, font_system: &mut FontSystem, action: Action);
    fn draw<F>(
        &self,
        font_system: &mut FontSystem,
        cache: &mut SwashCache,
        color: Color,
        f: F
    )
       where F: FnMut(i32, i32, u32, u32, Color);

    // Provided method
    fn borrow_with<'a>(
        &'a mut self,
        font_system: &'a mut FontSystem
    ) -> BorrowedWithFontSystem<'a, Self>
       where Self: Sized { ... }

}

A trait to allow easy replacements of Editor, like SyntaxEditor

Required Methods

fn buffer(&self) -> &Buffer

Get the internal Buffer

fn buffer_mut(&mut self) -> &mut Buffer

Get the internal Buffer, mutably

fn cursor(&self) -> Cursor

Get the current cursor

fn set_cursor(&mut self, cursor: Cursor)

Set the current cursor

fn select_opt(&self) -> Option<Cursor>

Get the current selection position

fn set_select_opt(&mut self, select_opt: Option<Cursor>)

Set the current selection position

fn shape_as_needed(&mut self, font_system: &mut FontSystem)

Shape lines until scroll, after adjusting scroll if the cursor moved

fn copy_selection(&mut self) -> Option<String>

Copy selection

fn delete_selection(&mut self) -> bool

Delete selection, adjusting cursor and returning true if there was a selection

fn insert_string(&mut self, data: &str, attrs_list: Option<AttrsList>)

Insert a string at the current cursor or replacing the current selection with the given attributes, or with the previous character’s attributes if None is given.

fn action(&mut self, font_system: &mut FontSystem, action: Action)

Perform an Action on the editor

fn draw<F>( &self, font_system: &mut FontSystem, cache: &mut SwashCache, color: Color, f: F )where F: FnMut(i32, i32, u32, u32, Color),

Draw the editor

Provided Methods

fn borrow_with<'a>( &'a mut self, font_system: &'a mut FontSystem ) -> BorrowedWithFontSystem<'a, Self>where Self: Sized,

Mutably borrows self together with an FontSystem for more convenient methods