Trait cosmic_text::Edit

pub trait Edit<'a> {
    fn buffer(&self) -> &Buffer<'a>;
    fn buffer_mut(&mut self) -> &mut Buffer<'a>;
    fn cursor(&self) -> Cursor;
    fn select_opt(&self) -> Option<Cursor>;
    fn set_select_opt(&mut self, select_opt: Option<Cursor>);
    fn shape_as_needed(&mut self);
    fn copy_selection(&mut self) -> Option<String>;
    fn delete_selection(&mut self) -> bool;
    fn action(&mut self, action: Action);
    fn draw<F>(&self, cache: &mut SwashCache<'_>, color: Color, f: F)
    where
        F: FnMut(i32, i32, u32, u32, Color);
}

A trait to allow easy replacements of Editor, like SyntaxEditor

Required Methods

fn buffer(&self) -> &Buffer<'a>

Get the internal Buffer

fn buffer_mut(&mut self) -> &mut Buffer<'a>

Get the internal Buffer, mutably

fn cursor(&self) -> Cursor

Get the current cursor position

fn select_opt(&self) -> Option<Cursor>

Get the current selection position

fn set_select_opt(&mut self, select_opt: Option<Cursor>)

Set the current selection position

fn shape_as_needed(&mut self)

Shape lines until scroll, after adjusting scroll if the cursor moved

fn copy_selection(&mut self) -> Option<String>

Copy selection

fn delete_selection(&mut self) -> bool

Delete selection, adjusting cursor and returning true if there was a selection

fn action(&mut self, action: Action)

Perform an Action on the editor

fn draw<F>(&self, cache: &mut SwashCache<'_>, color: Color, f: F)where
F: FnMut(i32, i32, u32, u32, Color),

Draw the editor